requirements.txt
asgiref==3.8.1
boto3==1.26.39
botocore==1.29.40
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-database-url==2.3.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.2.0
django-cors-headers==4.4.0
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
s3transfer==0.6.2
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
whitenoise==6.8.2
